Rename icons folder to images-and-icons. (#8)
