+++ src/components/images-and-icons/Icon/Icon.stories.tsx
