+++ src/components/images-and-icons/Icon/Icon.tsx
