+++ src/components/images-and-icons/Icon/icons-map.ts
