+++ src/components/index.ts
@@ -2,7 +2,7 @@ export * from './actions'
 export * from './base'
 export * from './feedback-indicators'
 export * from './forms'
-export * from './icons'
+export * from './images-and-icons'
 export * from './navigation'
 export * from './structure'
 export * from './text'
