+++ src/components/navigation/TopBar/Icon/Icon.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import { IconName } from '../../../icons'
+import { IconName } from '../../../images-and-icons'
 import { IconButton } from '../../../actions'
 
 export interface IconProps {
