+++ src/components/navigation/TopBar/TopBar.tsx
@@ -3,7 +3,7 @@ import { View } from 'react-native'
 import { useSafeArea } from 'react-native-safe-area-context'
 import { useStyles } from '../../../theme'
 import { shameStyles } from '../../../theme/shame-styles'
-import { IconName } from '../../icons/Icon/Icon'
+import { IconName } from '../../images-and-icons/Icon/Icon'
 import { Heading } from '../../text/Heading/Heading'
 import { Box } from '../../structure/Box/Box'
 import { IconButton } from '../../actions/IconButton/IconButton'
