+++ src/components/screens/Money.stories.tsx
@@ -6,7 +6,7 @@ import { DisplayText } from '../text/DisplayText/DisplayText'
 import { Divider } from '../structure/Divider/Divider'
 import { Card } from '../structure/Card/Card'
 import { BodyText } from '../text/BodyText/BodyText'
-import { IconName, Icon } from '../icons/Icon/Icon'
+import { IconName, Icon } from '../images-and-icons/Icon/Icon'
 import { getStoryTitle } from '../../storybook/get-story-title'
 import { PhoneScreen } from '../../storybook/decorators/PhoneScreen'
 import { Screen } from '../structure'
