+++ src/components/screens/Profile.stories.tsx
@@ -5,7 +5,7 @@ import { DisplayText } from '../text/DisplayText/DisplayText'
 import { Box } from '../structure/Box/Box'
 import { Divider } from '../structure/Divider/Divider'
 import { Touchable } from '../base/Touchable/Touchable'
-import { IconName, Icon } from '../icons/Icon/Icon'
+import { IconName, Icon } from '../images-and-icons/Icon/Icon'
 import { BodyText } from '../text/BodyText/BodyText'
 import { getStoryTitle } from '../../storybook/get-story-title'
 import { PhoneScreen } from '../../storybook/decorators/PhoneScreen'
