+++ src/components/structure/Card/Card.tsx
@@ -7,7 +7,7 @@ import { Box } from '../Box/Box'
 import { Divider } from '../Divider/Divider'
 import { Touchable } from '../../base/Touchable/Touchable'
 import { TextStyle } from '../../text/TextStyle/TextStyle'
-import { Icon } from '../../icons/Icon/Icon'
+import { Icon } from '../../images-and-icons/Icon/Icon'
 import { BodyText } from '../../text/BodyText/BodyText'
 import { shameStyles } from '../../../theme/shame-styles'
 import { TextAction } from '../../actions'
